Removed accountchooser madness
Added terrible back button to item page, not sure how to handle that yet.
Now that accountchooser doesn't happen when using the login with username and password we can correctly redirect the user to wherever they came from.
Added logged in user name and closing of side menu when mobile.